requirements.txt
streamlit
streamlit-option-menu
streamlit-image-comparison
pandas
google-generativeai
Pillow
PyMuPDF